iotivity/base.py: Added base iotivity test class
Add an iotivity base class for runtime tests, also moved
the setup and teardown funcionality of client.py to the base
class. Since many iotivity setup methods are similar, other
iotivity classes should use this base class too.
